Added getBerryEnergy, getBerryTitle, and completed getBerryInfo
getBerryEnergy returns an int value of the energy of the product in kcal
per 100g. -1 is used to represent that there is no known kcal value.
getBerryTitle returns a string containing the title of the product,
gained from a doc.
getBerryInfo has been completed and returns a Product object containing
all the needed information about the product.