Added Results, ProductSerializer and ProductToJSON classes
Results class is used to wrap the list of products up with other
information, ready for Json parsing.
ProductSerializer is a custom serializer for Gson, which is used to omit
kcal fields with a value of -1.
ProductToJson contains the method used to convert the result object into
a json string, using the serializer.